Graded-Ass/requirements.txt
streamlit>=1.32.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.2
lxml>=5.1.0
dspy-ai>=2.6.0
python-dotenv>=1.0.1
watchdog>=4.0.0
Pillow>=10.0.0
